refactor(Images): Split ito separate modules
