+++ rust/images/src/change_set.rs
@@ -1,0 +1,334 @@
+use std::{
+    collections::HashMap,
+    ffi::OsString,
+    io,
+    path::{Path, PathBuf},
+};
+
+use chrono::Utc;
+use eyre::{bail, eyre, Result};
+use oci_spec::image::{Descriptor, DescriptorBuilder, MediaType};
+
+use archive_utils::{flate2, tar, zstd};
+use hash_utils::{sha2::Digest, sha2::Sha256};
+
+use crate::blob_writer::BlobWriter;
+
+/// A change in a path between two snapshots
+///
+/// This enum represents the [Change Types](https://github.com/opencontainers/image-spec/blob/main/layer.md#change-types)
+/// described in the OCI spec.
+#[derive(Debug, PartialEq)]
+pub enum Change {
+    Added(String),
+    Modified(String),
+    Removed(String),
+}
+
+/// The set of changes between two snapshots
+///
+/// Represents the set of changes between two filesystem snapshots as described in
+/// [OCI Image Layer Filesystem Changeset](https://github.com/opencontainers/image-spec/blob/main/layer.md)
+pub struct ChangeSet {
+    /// The source directory, on the local filesystem, for the changes
+    source_dir: PathBuf,
+
+    /// The destination directory, within the image's root filesystem, for the changes
+    dest_dir: PathBuf,
+
+    /// The change items
+    pub(crate) items: Vec<Change>,
+}
+
+impl ChangeSet {
+    /// Create a new set of snapshot changes
+    pub fn new<P: AsRef<Path>>(source_dir: P, dest_dir: P, items: Vec<Change>) -> Self {
+        let source_dir = source_dir.as_ref().to_path_buf();
+
+        // Parths in tar archive must be relative so stri any leading slash
+        let dest_dir = dest_dir.as_ref().to_path_buf();
+        let dest_dir = match dest_dir.strip_prefix("/") {
+            Ok(dir) => dir.to_owned(),
+            Err(_) => dest_dir,
+        };
+
+        Self {
+            source_dir,
+            dest_dir,
+            items,
+        }
+    }
+
+    /// Creates an OCI layer for the set of changes
+    ///
+    /// This implements the [Representing Changes](https://github.com/opencontainers/image-spec/blob/main/layer.md#representing-changes)
+    /// section of the OCI image spec:
+    ///
+    /// - `Added` and `Modified` paths are added to the archive.
+    /// - `Removed` paths are represented as "whiteout" files.
+    ///
+    /// Note that two SHA256 hashes are calculated, one for the `DiffID` of a changeset (calculated in this function
+    /// and used in the image config file) and one for the digest which (calculated by the [`BlobWriter`] and used in the image manifest).
+    /// A useful diagram showing how these are calculated and used is available
+    /// [here](https://github.com/google/go-containerregistry/blob/main/pkg/v1/remote/README.md#anatomy-of-an-image-upload).
+    ///
+    /// # Arguments
+    ///
+    /// - `layout_dir`: the image directory to write the layer to (to the `blob/sha256` subdirectory)
+    pub fn write_layer<P: AsRef<Path>>(
+        self,
+        layout_dir: P,
+        media_type: &MediaType,
+    ) -> Result<(String, Descriptor)> {
+        if self.items.is_empty() {
+            return Ok((
+                "<empty>".to_string(),
+                DescriptorBuilder::default()
+                    .media_type(media_type.clone())
+                    .digest("<none>")
+                    .size(0)
+                    .build()?,
+            ));
+        }
+
+        tracing::info!(
+            "Writing image layer from changeset for `{}`",
+            self.source_dir.display()
+        );
+
+        let mut diffid_hash = Sha256::new();
+        let mut blob_writer = BlobWriter::new(&layout_dir, media_type.to_owned())?;
+
+        let changes = self.items.len();
+        let mut additions: Vec<String> = Vec::new();
+        let mut modifications: Vec<String> = Vec::new();
+        let mut deletions: Vec<String> = Vec::new();
+
+        {
+            enum LayerEncoder<'lt> {
+                Plain(&'lt mut BlobWriter),
+                Gzip(flate2::write::GzEncoder<&'lt mut BlobWriter>),
+                Zstd(zstd::stream::AutoFinishEncoder<'lt, &'lt mut BlobWriter>),
+            }
+
+            struct LayerWriter<'lt> {
+                diffid_hash: &'lt mut Sha256,
+                layer_encoder: LayerEncoder<'lt>,
+            }
+
+            impl<'lt> io::Write for LayerWriter<'lt> {
+                fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
+                    self.diffid_hash.update(buf);
+                    match &mut self.layer_encoder {
+                        LayerEncoder::Plain(encoder) => encoder.write_all(buf)?,
+                        LayerEncoder::Gzip(encoder) => encoder.write_all(buf)?,
+                        LayerEncoder::Zstd(encoder) => encoder.write_all(buf)?,
+                    }
+                    Ok(buf.len())
+                }
+
+                fn flush(&mut self) -> io::Result<()> {
+                    Ok(())
+                }
+            }
+
+            let layer_encoder = match media_type {
+                MediaType::ImageLayer => LayerEncoder::Plain(&mut blob_writer),
+                MediaType::ImageLayerGzip => LayerEncoder::Gzip(flate2::write::GzEncoder::new(
+                    &mut blob_writer,
+                    flate2::Compression::new(4),
+                )),
+                MediaType::ImageLayerZstd => LayerEncoder::Zstd(
+                    zstd::stream::Encoder::new(&mut blob_writer, 4)?.auto_finish(),
+                ),
+                _ => bail!("Unhandled media type for layer: {}", media_type.to_string()),
+            };
+
+            let mut layer_writer = LayerWriter {
+                diffid_hash: &mut diffid_hash,
+                layer_encoder,
+            };
+
+            let mut archive = tar::Builder::new(&mut layer_writer);
+
+            // Add an entry for the `dest_dir` (and any of its parent) so that ownership (and other
+            // metadata) of `source_dir` is maintained. If not done then there are issues with non-root
+            // users writing to the `workspace` and `layers` directories and  their subdirectories.
+            let mut path = PathBuf::new();
+            for part in self.dest_dir.components() {
+                path = path.join(part);
+                archive.append_path_with_name(&self.source_dir, &path)?;
+            }
+
+            // Add each change
+            for change in self.items {
+                match change {
+                    Change::Added(ref path) | Change::Modified(ref path) => {
+                        if let Err(error) = archive.append_path_with_name(
+                            self.source_dir.join(path),
+                            self.dest_dir.join(path),
+                        ) {
+                            tracing::debug!(
+                                "While adding added or modified entry to layer: {}",
+                                error
+                            )
+                        } else {
+                            match change {
+                                Change::Added(..) => additions.push(path.to_string()),
+                                Change::Modified(..) => modifications.push(path.to_string()),
+                                _ => unreachable!(),
+                            }
+                        }
+                    }
+                    Change::Removed(path) => {
+                        let path_buf = PathBuf::from(&path);
+                        let basename = path_buf
+                            .file_name()
+                            .ok_or_else(|| eyre!("Path has no file name"))?;
+                        let mut whiteout = OsString::from(".wh.".to_string());
+                        whiteout.push(basename);
+                        let path_buf = match path_buf.parent() {
+                            Some(parent) => parent.join(whiteout),
+                            None => PathBuf::from(whiteout),
+                        };
+                        let path_buf = self.dest_dir.join(path_buf);
+
+                        let mut header = tar::Header::new_gnu();
+                        header.set_path(path_buf)?;
+                        header.set_size(0);
+                        header.set_cksum();
+                        let data: &[u8] = &[];
+
+                        if let Err(error) = archive.append(&header, data) {
+                            tracing::debug!("While adding whiteout entry for file: {}", error)
+                        } else {
+                            deletions.push(path)
+                        }
+                    }
+                };
+            }
+        }
+
+        let diff_id = format!("sha256:{:x}", diffid_hash.finalize());
+
+        let mut annotations: HashMap<String, String> = [
+            ("io.stencila.version", env!("CARGO_PKG_VERSION").to_string()),
+            ("io.stencila.layer.created", Utc::now().to_rfc3339()),
+            (
+                "io.stencila.layer.directory",
+                self.dest_dir.to_string_lossy().to_string(),
+            ),
+            ("io.stencila.layer.changes", changes.to_string()),
+        ]
+        .map(|(name, value)| (name.to_string(), value))
+        .into();
+
+        fn first_100(vec: Vec<String>) -> String {
+            vec[..(std::cmp::min(vec.len(), 100))].join(":")
+        }
+        if !additions.is_empty() {
+            annotations.insert(
+                "io.stencila.layer.additions".to_string(),
+                first_100(additions),
+            );
+        }
+        if !modifications.is_empty() {
+            annotations.insert(
+                "io.stencila.layer.modifications".to_string(),
+                first_100(modifications),
+            );
+        }
+        if !deletions.is_empty() {
+            annotations.insert(
+                "io.stencila.layer.deletions".to_string(),
+                first_100(deletions),
+            );
+        }
+
+        let descriptor = blob_writer.finish(Some(annotations))?;
+
+        Ok((diff_id, descriptor))
+    }
+
+    /// Get the path of a layer blob within an image directory
+    ///
+    /// # Arguments
+    ///
+    /// - `image_dir`: the image directory
+    /// - `digest`: the digest of the layer (with or without the "sha256:" prefix)
+    fn layer_path<P: AsRef<Path>>(image_dir: P, digest: &str) -> PathBuf {
+        image_dir
+            .as_ref()
+            .join("blobs")
+            .join("sha256")
+            .join(digest.strip_prefix("sha256:").unwrap_or(digest))
+    }
+
+    /// Read a layer blob (a compressed tar archive) from an image directory
+    ///
+    /// At this stage, mainly just used for testing.
+    ///
+    /// # Arguments
+    ///
+    /// - `image_dir`: the image directory
+    /// - `digest`: the digest of the layer (with or without the "sha256:" prefix)
+    pub fn read_layer<P: AsRef<Path>>(
+        image_dir: P,
+        digest: &str,
+    ) -> Result<tar::Archive<flate2::read::GzDecoder<std::fs::File>>> {
+        use std::fs;
+
+        let path = Self::layer_path(image_dir, digest);
+        let file = fs::File::open(&path)?;
+        let decoder = flate2::read::GzDecoder::new(file);
+        let archive = tar::Archive::new(decoder);
+        Ok(archive)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use hash_utils::file_sha256_hex;
+    use test_utils::tempfile::tempdir;
+
+    use crate::snapshot::Snapshot;
+
+    use super::*;
+
+    /// Test that the descriptor for a layer is accurate (SHA256 and size are same as
+    /// when independently calculated)
+    #[test]
+    fn changes_layer() -> Result<()> {
+        use std::fs;
+
+        let source_dir = tempdir()?;
+        let image_dir = tempdir()?;
+
+        let snap = Snapshot::new(&source_dir, "workspace");
+
+        fs::write(&source_dir.path().join("some-file.txt"), "Hello")?;
+
+        // Create a layer archive, diffid and descriptor
+
+        let changes = snap.changes();
+        let (diff_id, descriptor) = changes.write_layer(&image_dir, &MediaType::ImageLayerGzip)?;
+
+        assert_eq!(diff_id.len(), 7 + 64);
+        assert!(diff_id.starts_with("sha256:"));
+
+        // Test that size and digest in the descriptor is as for the file
+        let archive = image_dir
+            .path()
+            .join("blobs")
+            .join("sha256")
+            .join(descriptor.digest().strip_prefix("sha256:").unwrap());
+
+        let size = fs::metadata(&archive)?.len() as i64;
+        assert_eq!(descriptor.size(), size);
+
+        let digest = format!("sha256:{}", file_sha256_hex(&archive)?);
+        assert_eq!(descriptor.digest(), &digest);
+
+        Ok(())
+    }
+}
