+++ rust/images/src/image.rs
@@ -1,43 +1,25 @@
 use std::{
     collections::HashMap,
     env,
-    ffi::OsString,
-    hash::Hasher,
-    io,
-    os::unix::prelude::MetadataExt,
     path::{Path, PathBuf},
     str::FromStr,
-    time::{SystemTime, UNIX_EPOCH},
 };
 
 use chrono::Utc;
 use eyre::{bail, eyre, Result};
-use jwalk::WalkDirGeneric;
 use oci_spec::image::{
-    Descriptor, DescriptorBuilder, History, HistoryBuilder, ImageConfiguration,
-    ImageConfigurationBuilder, ImageIndexBuilder, ImageManifestBuilder, MediaType, RootFsBuilder,
-    SCHEMA_VERSION,
+    Descriptor, History, HistoryBuilder, ImageConfiguration, ImageConfigurationBuilder,
+    ImageIndexBuilder, ImageManifestBuilder, MediaType, RootFsBuilder, SCHEMA_VERSION,
 };
-use seahash::SeaHasher;
 
-// Serialization framework defaults to `rkyv` with fallback to `serde` JSON
-
-#[cfg(feature = "rkyv")]
-use rkyv::{Archive, Deserialize, Serialize};
-
-#[cfg(feature = "rkyv-safe")]
-use bytecheck::CheckBytes;
-
-#[cfg(not(feature = "rkyv"))]
-use serde::{Deserialize, Serialize};
-
-use archive_utils::{flate2, tar, zstd};
-use hash_utils::{sha2::Digest, sha2::Sha256, str_sha256_hex};
+use hash_utils::str_sha256_hex;
 use http_utils::tempfile::{tempdir, TempDir};
 
 use crate::{
+    blob_writer::BlobWriter,
     distribution::{Client, DOCKER_REGISTRY},
     media_types::ToDockerV2S2,
+    snapshot::Snapshot,
     utils::unique_string,
 };
 
@@ -157,7 +139,7 @@ pub struct Image {
     /// The working directory to build an image for
     ///
     /// Buildpacks will build layers based on the source code within this directory. Usually
-    /// the home directroy of a project. Defaults to the current directory.
+    /// the home directory of a project. Defaults to the current directory.
     working_dir: Option<PathBuf>,
 
     /// The image reference for this image
@@ -252,7 +234,7 @@ impl Image {
         let base = base
             .map(String::from)
             .or_else(|| std::env::var("STENCILA_IMAGE_REF").ok())
-            .unwrap_or_else(|| "stencila/stencila:nano".to_string())
+            .unwrap_or_else(|| "scratch".to_string())
             .parse()?;
 
         let layers_dir = layers_dir
@@ -678,707 +660,9 @@ impl Image {
     }
 }
 
-/// The set of changes between two snapshots
-///
-/// Represents the set of changes between two filesystem snapshots as described in
-/// [OCI Image Layer Filesystem Changeset](https://github.com/opencontainers/image-spec/blob/main/layer.md)
-struct ChangeSet {
-    /// The source directory, on the local filesystem, for the changes
-    source_dir: PathBuf,
-
-    /// The destination directory, within the image's root filesystem, for the changes
-    dest_dir: PathBuf,
-
-    /// The change items
-    items: Vec<Change>,
-}
-
-impl ChangeSet {
-    /// Create a new set of snapshot changes
-    fn new<P: AsRef<Path>>(source_dir: P, dest_dir: P, items: Vec<Change>) -> Self {
-        let source_dir = source_dir.as_ref().to_path_buf();
-
-        // Parths in tar archive must be relative so stri any leading slash
-        let dest_dir = dest_dir.as_ref().to_path_buf();
-        let dest_dir = match dest_dir.strip_prefix("/") {
-            Ok(dir) => dir.to_owned(),
-            Err(_) => dest_dir,
-        };
-
-        Self {
-            source_dir,
-            dest_dir,
-            items,
-        }
-    }
-
-    /// Get the number of changes in this set
-    fn len(&self) -> usize {
-        self.items.len()
-    }
-
-    /// Creates an OCI layer for the set of changes
-    ///
-    /// This implements the [Representing Changes](https://github.com/opencontainers/image-spec/blob/main/layer.md#representing-changes)
-    /// section of the OCI image spec:
-    ///
-    /// - `Added` and `Modified` paths are added to the archive.
-    /// - `Removed` paths are represented as "whiteout" files.
-    ///
-    /// Note that two SHA256 hashes are calculated, one for the `DiffID` of a changeset (calculated in this function
-    /// and used in the image config file) and one for the digest which (calculated by the [`BlobWriter`] and used in the image manifest).
-    /// A useful diagram showing how these are calculated and used is available
-    /// [here](https://github.com/google/go-containerregistry/blob/main/pkg/v1/remote/README.md#anatomy-of-an-image-upload).
-    ///
-    /// # Arguments
-    ///
-    /// - `layout_dir`: the image directory to write the layer to (to the `blob/sha256` subdirectory)
-    fn write_layer<P: AsRef<Path>>(
-        self,
-        layout_dir: P,
-        media_type: &MediaType,
-    ) -> Result<(String, Descriptor)> {
-        if self.len() == 0 {
-            return Ok((
-                "<empty>".to_string(),
-                DescriptorBuilder::default()
-                    .media_type(media_type.clone())
-                    .digest("<none>")
-                    .size(0)
-                    .build()?,
-            ));
-        }
-
-        tracing::info!(
-            "Writing image layer from changeset for `{}`",
-            self.source_dir.display()
-        );
-
-        let mut diffid_hash = Sha256::new();
-        let mut blob_writer = BlobWriter::new(&layout_dir, media_type.to_owned())?;
-
-        let changes = self.len();
-        let mut additions: Vec<String> = Vec::new();
-        let mut modifications: Vec<String> = Vec::new();
-        let mut deletions: Vec<String> = Vec::new();
-
-        {
-            enum LayerEncoder<'lt> {
-                Plain(&'lt mut BlobWriter),
-                Gzip(flate2::write::GzEncoder<&'lt mut BlobWriter>),
-                Zstd(zstd::stream::AutoFinishEncoder<'lt, &'lt mut BlobWriter>),
-            }
-
-            struct LayerWriter<'lt> {
-                diffid_hash: &'lt mut Sha256,
-                layer_encoder: LayerEncoder<'lt>,
-            }
-
-            impl<'lt> io::Write for LayerWriter<'lt> {
-                fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-                    self.diffid_hash.update(buf);
-                    match &mut self.layer_encoder {
-                        LayerEncoder::Plain(encoder) => encoder.write_all(buf)?,
-                        LayerEncoder::Gzip(encoder) => encoder.write_all(buf)?,
-                        LayerEncoder::Zstd(encoder) => encoder.write_all(buf)?,
-                    }
-                    Ok(buf.len())
-                }
-
-                fn flush(&mut self) -> io::Result<()> {
-                    Ok(())
-                }
-            }
-
-            let layer_encoder = match media_type {
-                MediaType::ImageLayer => LayerEncoder::Plain(&mut blob_writer),
-                MediaType::ImageLayerGzip => LayerEncoder::Gzip(flate2::write::GzEncoder::new(
-                    &mut blob_writer,
-                    flate2::Compression::new(4),
-                )),
-                MediaType::ImageLayerZstd => LayerEncoder::Zstd(
-                    zstd::stream::Encoder::new(&mut blob_writer, 4)?.auto_finish(),
-                ),
-                _ => bail!("Unhandled media type for layer: {}", media_type.to_string()),
-            };
-
-            let mut layer_writer = LayerWriter {
-                diffid_hash: &mut diffid_hash,
-                layer_encoder,
-            };
-
-            let mut archive = tar::Builder::new(&mut layer_writer);
-
-            // Add an entry for the `dest_dir` so that ownership and other metadata is maintained
-            archive.append_path_with_name(&self.source_dir, &self.dest_dir)?;
-
-            // Add each change
-            for change in self.items {
-                match change {
-                    Change::Added(ref path) | Change::Modified(ref path) => {
-                        if let Err(error) = archive.append_path_with_name(
-                            self.source_dir.join(path),
-                            self.dest_dir.join(path),
-                        ) {
-                            tracing::debug!(
-                                "While adding added or modified entry to layer: {}",
-                                error
-                            )
-                        } else {
-                            match change {
-                                Change::Added(..) => additions.push(path.to_string()),
-                                Change::Modified(..) => modifications.push(path.to_string()),
-                                _ => unreachable!(),
-                            }
-                        }
-                    }
-                    Change::Removed(path) => {
-                        let path_buf = PathBuf::from(&path);
-                        let basename = path_buf
-                            .file_name()
-                            .ok_or_else(|| eyre!("Path has no file name"))?;
-                        let mut whiteout = OsString::from(".wh.".to_string());
-                        whiteout.push(basename);
-                        let path_buf = match path_buf.parent() {
-                            Some(parent) => parent.join(whiteout),
-                            None => PathBuf::from(whiteout),
-                        };
-                        let path_buf = self.dest_dir.join(path_buf);
-
-                        let mut header = tar::Header::new_gnu();
-                        header.set_path(path_buf)?;
-                        header.set_size(0);
-                        header.set_cksum();
-                        let data: &[u8] = &[];
-
-                        if let Err(error) = archive.append(&header, data) {
-                            tracing::debug!("While adding whiteout entry for file: {}", error)
-                        } else {
-                            deletions.push(path)
-                        }
-                    }
-                };
-            }
-        }
-
-        let diff_id = format!("sha256:{:x}", diffid_hash.finalize());
-
-        let mut annotations: HashMap<String, String> = [
-            ("io.stencila.version", env!("CARGO_PKG_VERSION").to_string()),
-            ("io.stencila.layer.created", Utc::now().to_rfc3339()),
-            (
-                "io.stencila.layer.directory",
-                self.dest_dir.to_string_lossy().to_string(),
-            ),
-            ("io.stencila.layer.changes", changes.to_string()),
-        ]
-        .map(|(name, value)| (name.to_string(), value))
-        .into();
-
-        fn first_100(vec: Vec<String>) -> String {
-            vec[..(std::cmp::min(vec.len(), 100))].join(":")
-        }
-        if !additions.is_empty() {
-            annotations.insert(
-                "io.stencila.layer.additions".to_string(),
-                first_100(additions),
-            );
-        }
-        if !modifications.is_empty() {
-            annotations.insert(
-                "io.stencila.layer.modifications".to_string(),
-                first_100(modifications),
-            );
-        }
-        if !deletions.is_empty() {
-            annotations.insert(
-                "io.stencila.layer.deletions".to_string(),
-                first_100(deletions),
-            );
-        }
-
-        let descriptor = blob_writer.finish(Some(annotations))?;
-
-        Ok((diff_id, descriptor))
-    }
-
-    /// Get the path of a layer blob within an image directory
-    ///
-    /// # Arguments
-    ///
-    /// - `image_dir`: the image directory
-    /// - `digest`: the digest of the layer (with or without the "sha256:" prefix)
-    fn layer_path<P: AsRef<Path>>(image_dir: P, digest: &str) -> PathBuf {
-        image_dir
-            .as_ref()
-            .join("blobs")
-            .join("sha256")
-            .join(digest.strip_prefix("sha256:").unwrap_or(digest))
-    }
-
-    /// Read a layer blob (a compressed tar archive) from an image directory
-    ///
-    /// At this stage, mainly just used for testing.
-    ///
-    /// # Arguments
-    ///
-    /// - `image_dir`: the image directory
-    /// - `digest`: the digest of the layer (with or without the "sha256:" prefix)
-    fn read_layer<P: AsRef<Path>>(
-        image_dir: P,
-        digest: &str,
-    ) -> Result<tar::Archive<flate2::read::GzDecoder<std::fs::File>>> {
-        use std::fs;
-
-        let path = Self::layer_path(image_dir, digest);
-        let file = fs::File::open(&path)?;
-        let decoder = flate2::read::GzDecoder::new(file);
-        let archive = tar::Archive::new(decoder);
-        Ok(archive)
-    }
-}
-
-/// A change in a path between two snapshots
-///
-/// This enum represents the [Change Types](https://github.com/opencontainers/image-spec/blob/main/layer.md#change-types)
-/// described in the OCI spec.
-#[derive(Debug, PartialEq)]
-enum Change {
-    Added(String),
-    Modified(String),
-    Removed(String),
-}
-
-/// A snapshot of the files and directories in a directory
-///
-/// A snapshot is created at the start of a session and stored to disk. Another snapshot
-/// is taken at the end of session. The changes between the snapshots are used to create
-/// an image layer.
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
-#[cfg_attr(feature = "rkyv", derive(Archive))]
-#[cfg_attr(feature = "rkyv-safe", archive_attr(derive(CheckBytes)))]
-struct Snapshot {
-    /// The source directory, on the local filesystem, for the snapshot
-    source_dir: String,
-
-    /// The destination directory, on the image's root filesystem, for the snapshot
-    dest_dir: String,
-
-    /// Entries in the snapshot
-    entries: HashMap<String, SnapshotEntry>,
-}
-
-impl Snapshot {
-    /// Create a new snapshot of a directory
-    ///
-    /// If there is a `.dockerignore` or `.containerignore` file in source directory then it will be
-    /// used to exclude paths, including those in child sub-directories.
-    fn new<S: AsRef<Path>, D: AsRef<Path>>(source_dir: S, dest_dir: D) -> Self {
-        let source_dir = source_dir.as_ref().to_path_buf();
-        let dest_dir = dest_dir.as_ref().to_path_buf();
-
-        let docker_ignore = source_dir.join(".dockerignore");
-        let container_ignore = source_dir.join(".containerignore");
-        fn parse_ignore_file(path: &Path) -> Option<gitignore::File> {
-            match gitignore::File::new(&path) {
-                Ok(file) => Some(file),
-                Err(error) => {
-                    tracing::warn!(
-                        "Error while parsing `{}`; will not be used: {}",
-                        path.display(),
-                        error
-                    );
-                    None
-                }
-            }
-        }
-        let ignore_file = if docker_ignore.exists() {
-            parse_ignore_file(&docker_ignore)
-        } else if container_ignore.exists() {
-            parse_ignore_file(&container_ignore)
-        } else {
-            None
-        };
-
-        let entries = WalkDirGeneric::<((), SnapshotEntry)>::new(&source_dir)
-            .skip_hidden(false)
-            .process_read_dir(|_depth, _path, _read_dir_state, children| {
-                children.iter_mut().flatten().for_each(|dir_entry| {
-                    if !dir_entry.file_type.is_dir() {
-                        dir_entry.client_state = SnapshotEntry::new(
-                            &dir_entry.path(),
-                            &dir_entry.file_type(),
-                            dir_entry.metadata().ok(),
-                        );
-                    }
-                })
-            })
-            .into_iter()
-            .filter_map(|entry_result| match entry_result {
-                Ok(entry) => {
-                    let path = entry.path();
-
-                    // Check that the file should not be ignored
-                    if let Some(true) = ignore_file
-                        .as_ref()
-                        .map(|ignore_file| ignore_file.is_excluded(&path).ok())
-                        .flatten()
-                    {
-                        return None;
-                    };
-
-                    let relative_path = path
-                        .strip_prefix(&source_dir)
-                        .expect("Should always be able to strip the root dir");
-                    match relative_path == PathBuf::from("") {
-                        true => None, // This is the entry for the dir itself so ignore it
-                        false => Some((
-                            relative_path.to_string_lossy().to_string(), // Should be lossless on Linux (and MacOS)
-                            entry.client_state,
-                        )),
-                    }
-                }
-                Err(error) => {
-                    tracing::error!("While snapshotting `{}`: {}", source_dir.display(), error);
-                    None
-                }
-            })
-            .collect();
-
-        let source_dir = source_dir.to_string_lossy().to_string();
-        let dest_dir = dest_dir.to_string_lossy().to_string();
-        Self {
-            source_dir,
-            dest_dir,
-            entries,
-        }
-    }
-
-    /// Create a new snapshot by repeating the current one
-    fn repeat(&self) -> Self {
-        Self::new(&self.source_dir, &self.dest_dir)
-    }
-
-    /// Write a snapshot to a file
-    fn write<P: AsRef<Path>>(&self, path: P) -> Result<()> {
-        use std::fs;
-
-        #[cfg(feature = "rkyv")]
-        {
-            let bytes = rkyv::to_bytes::<Self, 256>(self)?;
-            fs::write(path, bytes)?;
-        }
-
-        #[cfg(not(feature = "rkyv"))]
-        {
-            let json = serde_json::to_string_pretty(self)?;
-            fs::write(path, json)?;
-        }
-
-        Ok(())
-    }
-
-    /// Read a snapshot from a file
-    fn read<P: AsRef<Path>>(path: P) -> Result<Self> {
-        use std::fs;
-
-        #[cfg(feature = "rkyv")]
-        {
-            let bytes = fs::read(path)?;
-
-            #[cfg(feature = "rkyv-safe")]
-            let archived = match rkyv::check_archived_root::<Self>(&bytes[..]) {
-                Ok(archived) => archived,
-                Err(error) => {
-                    bail!("While checking archive: {}", error)
-                }
-            };
-
-            #[cfg(not(feature = "rkyv-safe"))]
-            let archived = unsafe { rkyv::archived_root::<Self>(&bytes[..]) };
-
-            let snapshot = archived.deserialize(&mut rkyv::Infallible)?;
-            Ok(snapshot)
-        }
-
-        #[cfg(not(feature = "rkyv"))]
-        {
-            let json = fs::read_to_string(&path)?;
-            let snapshot = serde_json::from_str(&json)?;
-            Ok(snapshot)
-        }
-    }
-
-    /// Create a set of changes that replicate the current snapshot using only additions
-    fn replicate(&self) -> ChangeSet {
-        let changes = self
-            .entries
-            .keys()
-            .map(|path| Change::Added(path.into()))
-            .collect();
-        ChangeSet::new(&self.source_dir, &self.dest_dir, changes)
-    }
-
-    /// Create a set of changes by determining the difference between two snapshots
-    fn diff(&self, other: &Snapshot) -> ChangeSet {
-        let mut changes = Vec::new();
-        for (path, entry) in self.entries.iter() {
-            match other.entries.get(path) {
-                Some(other_entry) => {
-                    if entry != other_entry {
-                        changes.push(Change::Modified(path.into()))
-                    }
-                }
-                None => changes.push(Change::Removed(path.into())),
-            }
-        }
-        for path in other.entries.keys() {
-            if !self.entries.contains_key(path) {
-                changes.push(Change::Added(path.into()))
-            }
-        }
-        ChangeSet::new(&self.source_dir, &self.dest_dir, changes)
-    }
-
-    /// Create a set of changes by repeating the current snapshot
-    ///
-    /// Convenience function for combining calls to `repeat` and `diff.
-    fn changes(&self) -> ChangeSet {
-        self.diff(&self.repeat())
-    }
-
-    /// Create a layer by repeating the current snapshot
-    ///
-    /// # Arguments
-    ///
-    /// - `diff`: Whether to create the layer as the difference to the original snapshot
-    ///           (the usual) or as a replicate.
-    fn write_layer<P: AsRef<Path>>(
-        &self,
-        layout_dir: P,
-        diff: bool,
-        media_type: &MediaType,
-    ) -> Result<(String, Descriptor)> {
-        let new = self.repeat();
-        let changeset = if diff {
-            self.diff(&new)
-        } else {
-            new.replicate()
-        };
-        changeset.write_layer(layout_dir, media_type)
-    }
-}
-
-/// An entry for a file or directory in a snapshot
-#[derive(Debug, Default, PartialEq, Serialize, Deserialize)]
-#[cfg_attr(feature = "rkyv", derive(Archive))]
-#[cfg_attr(feature = "rkyv-safe", archive_attr(derive(CheckBytes)))]
-struct SnapshotEntry {
-    /// Metadata on the file or directory
-    ///
-    /// Should only be `None` if there was an error getting the metadata
-    /// while creating the snapshot.
-    metadata: Option<SnapshotEntryMetadata>,
-
-    /// Hash of the content of the file
-    ///
-    /// Will be `None` if the entry is a directory
-    fingerprint: Option<u64>,
-}
-
-/// Filesystem metadata for a snapshot entry
-///
-/// Only includes the metadata that needs to be differences. For that reason,
-/// does not record `modified` time since that would create a false positive
-/// difference (if all other attributes were the same).
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
-#[cfg_attr(feature = "rkyv", derive(Archive))]
-#[cfg_attr(feature = "rkyv-safe", archive_attr(derive(CheckBytes)))]
-struct SnapshotEntryMetadata {
-    uid: u32,
-    gid: u32,
-    readonly: bool,
-}
-
-impl SnapshotEntry {
-    /// Create a new snapshot entry
-    fn new(
-        path: &Path,
-        file_type: &std::fs::FileType,
-        metadata: Option<std::fs::Metadata>,
-    ) -> Self {
-        let metadata = metadata.map(|metadata| SnapshotEntryMetadata {
-            uid: metadata.uid(),
-            gid: metadata.gid(),
-            readonly: metadata.permissions().readonly(),
-        });
-
-        let fingerprint = if file_type.is_file() {
-            match Self::file_fingerprint::<SeaHasher>(path) {
-                Ok(fingerprint) => Some(fingerprint),
-                Err(error) => {
-                    tracing::error!("While fingerprinting file `{}`: {}", path.display(), error);
-                    None
-                }
-            }
-        } else {
-            None
-        };
-
-        Self {
-            metadata,
-            fingerprint,
-        }
-    }
-
-    /// Generate a hash of a file's content
-    ///
-    /// Used to generate a fingerprint
-    ///
-    /// Based on https://github.com/jRimbault/yadf/blob/04205a57882ffa7d6a9ca05016e18214a38079b6/src/fs/hash.rs#L29
-    fn file_fingerprint<H>(path: &Path) -> io::Result<u64>
-    where
-        H: Hasher + Default,
-    {
-        struct HashWriter<H>(H);
-        impl<H: Hasher> io::Write for HashWriter<H> {
-            fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-                self.0.write(buf);
-                Ok(buf.len())
-            }
-
-            fn flush(&mut self) -> io::Result<()> {
-                Ok(())
-            }
-        }
-
-        let mut hasher = HashWriter(H::default());
-        io::copy(&mut std::fs::File::open(path)?, &mut hasher)?;
-        Ok(hasher.0.finish())
-    }
-
-    /// Get a timestamp from a file's created or modified system time
-    fn file_timestamp(time: Result<SystemTime, io::Error>) -> Option<u64> {
-        time.map(|system_time| {
-            system_time
-                .duration_since(UNIX_EPOCH)
-                .expect("Time should not go backwards")
-                .as_secs()
-        })
-        .ok()
-    }
-}
-
-/// A writer that calculates the size and SHA256 hash of files as they are written
-///
-/// Writes blobs into the `blobs/sha256` subdirectory of an image directory and returns
-/// an [OCI Content Descriptor](https://github.com/opencontainers/image-spec/blob/main/descriptor.md).
-///
-/// Allows use to do a single pass when writing files instead of reading them after writing in order
-/// to generate the SHA256 signature.
-struct BlobWriter {
-    /// The path to the `blobs/sha256` subdirectory where the blob is written to
-    blobs_dir: PathBuf,
-
-    /// The media type of the blob
-    media_type: MediaType,
-
-    /// The temporary file name of the blob (used before we know its final name, which is its SHA256 checksum)
-    file_name: PathBuf,
-
-    /// The file the blob is written to
-    file: std::fs::File,
-
-    /// The number of bytes in the blob content
-    bytes: usize,
-
-    /// The SHA256 hash of the blob content
-    hash: Sha256,
-}
-
-impl BlobWriter {
-    /// Create a new blob writer
-    ///
-    /// # Arguments
-    ///
-    /// - `image_dir`: the image directory (blobs are written to the `blobs/sha256` subdirectory of this)
-    /// - `media_type`: the media type of the blob
-    fn new<P: AsRef<Path>>(image_dir: P, media_type: MediaType) -> Result<Self> {
-        use std::fs::{self, File};
-
-        let blobs_dir = image_dir.as_ref().join("blobs").join("sha256");
-        fs::create_dir_all(&blobs_dir)?;
-
-        let filename = PathBuf::from(format!("temporary-{}", unique_string()));
-        let file = File::create(blobs_dir.join(&filename))?;
-
-        Ok(Self {
-            blobs_dir,
-            media_type,
-            file_name: filename,
-            file,
-            bytes: 0,
-            hash: Sha256::new(),
-        })
-    }
-
-    /// Finish writing the blob
-    ///
-    /// Finalizes the SHA256 hash, renames the file to the hex digest of that hash,
-    /// and returns a descriptor of the blob.
-    fn finish(self, annotations: Option<HashMap<String, String>>) -> Result<Descriptor> {
-        use std::fs;
-
-        let sha256 = format!("{:x}", self.hash.finalize());
-
-        fs::rename(
-            self.blobs_dir.join(self.file_name),
-            self.blobs_dir.join(&sha256),
-        )?;
-
-        let mut descriptor = DescriptorBuilder::default()
-            .media_type(self.media_type)
-            .size(self.bytes as i64)
-            .digest(format!("sha256:{}", sha256));
-        if let Some(annotations) = annotations {
-            descriptor = descriptor.annotations(annotations)
-        }
-        let descriptor = descriptor.build()?;
-
-        Ok(descriptor)
-    }
-
-    /// Write an object as a JSON based media type
-    fn write_json<P: AsRef<Path>, S: serde::Serialize>(
-        path: P,
-        media_type: MediaType,
-        object: &S,
-        annotations: Option<HashMap<String, String>>,
-    ) -> Result<Descriptor> {
-        let mut writer = Self::new(path, media_type)?;
-        serde_json::to_writer_pretty(&mut writer, object)?;
-        writer.finish(annotations)
-    }
-}
-
-impl io::Write for BlobWriter {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        self.file.write_all(buf)?;
-        self.bytes += buf.len();
-        self.hash.update(buf);
-        Ok(buf.len())
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        Ok(())
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use hash_utils::file_sha256_hex;
-    use test_snaps::fixtures;
-    use test_utils::{print_logs, tempfile::tempdir};
+    use test_utils::tempfile::tempdir;
 
     use super::*;
 
@@ -1438,196 +722,6 @@ mod tests {
         Ok(())
     }
 
-    /// Test that snapshots are correctly written to and read back from disk
-    #[test]
-    fn snapshot_serialization() -> Result<()> {
-        let working_dir = fixtures().join("projects").join("apt");
-
-        let temp = tempdir()?;
-        let snapshot_path = temp.path().join("test.snap");
-        let snapshot1 = Snapshot::new(working_dir, "/workspace");
-
-        snapshot1.write(&snapshot_path)?;
-
-        let snapshot2 = Snapshot::read(&snapshot_path)?;
-        assert_eq!(snapshot1, snapshot2);
-
-        Ok(())
-    }
-
-    /// Test that snap-shotting takes into consideration .dockerignore and .containerignore files
-    #[test]
-    fn snapshot_ignores() -> Result<()> {
-        use std::fs;
-
-        let temp = tempdir()?;
-        let source_dir = temp.path();
-
-        fs::write(source_dir.join("a.txt"), "A")?;
-        assert!(Snapshot::new(source_dir, "").entries.contains_key("a.txt"));
-
-        fs::write(source_dir.join(".dockerignore"), "*\n")?;
-        assert!(!Snapshot::new(source_dir, "").entries.contains_key("a.txt"));
-
-        fs::remove_file(source_dir.join(".dockerignore"))?;
-        fs::write(source_dir.join(".containerignore"), "*.txt\n")?;
-        assert!(!Snapshot::new(source_dir, "").entries.contains_key("a.txt"));
-
-        fs::remove_file(source_dir.join(".containerignore"))?;
-        fs::write(source_dir.join("b.txt"), "B")?;
-        fs::write(source_dir.join(".dockerignore"), "!a.txt\n")?;
-        let snapshot = Snapshot::new(source_dir, "");
-        assert!(snapshot.entries.contains_key("a.txt"));
-        assert!(snapshot.entries.contains_key("b.txt"));
-
-        Ok(())
-    }
-
-    /// Test snap-shotting, calculation of changesets, and the generation of layers from them.
-    #[test]
-    fn snapshot_changes() -> Result<()> {
-        use std::fs;
-
-        print_logs();
-
-        // Create a temporary directory as a text fixture and a tar file for writing / reading layers
-
-        let source_dir = tempdir()?;
-        let dest_dir = PathBuf::from("workspace");
-        let image_dir = tempdir()?;
-
-        // Create an initial snapshot which should be empty and has no changes with self
-
-        let snap1 = Snapshot::new(source_dir.path(), &dest_dir);
-        assert_eq!(snap1.entries.len(), 0);
-
-        let changes = snap1.diff(&snap1);
-        assert_eq!(changes.len(), 0);
-
-        // Add a file, create a new snapshot and check it has one entry and produces a change set
-        // with `Added` and tar has entry for it
-
-        let a_txt = "a.txt".to_string();
-        fs::write(source_dir.path().join(&a_txt), "Hello from a.txt")?;
-
-        let snap2 = snap1.repeat();
-        assert_eq!(snap2.entries.len(), 1);
-        assert_eq!(snap2.entries[&a_txt].fingerprint, Some(3958791156379554752));
-
-        let changes = snap1.diff(&snap2);
-        assert_eq!(changes.len(), 1);
-        assert_eq!(changes.items[0], Change::Added(a_txt.clone()));
-
-        let (.., descriptor) = changes.write_layer(&image_dir, &MediaType::ImageLayerGzip)?;
-
-        let mut layer = ChangeSet::read_layer(&image_dir, descriptor.digest())?;
-        let mut entries = layer.entries()?;
-        let entry = entries
-            .next()
-            .ok_or_else(|| eyre!("No entries in tar archive"))??;
-        assert_eq!(entry.path()?, dest_dir.join(&a_txt));
-        assert_eq!(entry.size(), 16);
-
-        // Repeat
-
-        let b_txt = "b.txt".to_string();
-        fs::write(source_dir.path().join(&b_txt), "Hello from b.txt")?;
-
-        let snap3 = snap1.repeat();
-        assert_eq!(snap3.entries.len(), 2);
-        assert_eq!(snap2.entries[&a_txt].fingerprint, Some(3958791156379554752));
-        assert_eq!(
-            snap3.entries[&b_txt].fingerprint,
-            Some(15548480638800185371)
-        );
-
-        let changes = snap2.diff(&snap3);
-        assert_eq!(changes.len(), 1);
-        assert_eq!(changes.items[0], Change::Added(b_txt.clone()));
-
-        // Remove a.txt and check that the change set has a `Removed` and tar has
-        // a whiteout entry of size 0
-
-        fs::remove_file(source_dir.path().join(&a_txt))?;
-
-        let snap4 = snap1.repeat();
-        assert_eq!(snap4.entries.len(), 1);
-        assert_eq!(
-            snap4.entries[&b_txt].fingerprint,
-            Some(15548480638800185371)
-        );
-
-        let changes = snap3.diff(&snap4);
-        assert_eq!(changes.len(), 1);
-        assert_eq!(changes.items[0], Change::Removed(a_txt));
-
-        let (.., descriptor) = changes.write_layer(&image_dir, &MediaType::ImageLayerGzip)?;
-        let mut layer = ChangeSet::read_layer(&image_dir, descriptor.digest())?;
-        let mut entries = layer.entries()?;
-        let entry = entries.next().unwrap()?;
-        assert_eq!(entry.path()?, dest_dir.join(".wh.a.txt"));
-        assert_eq!(entry.size(), 0);
-
-        // Modify b.txt and check that the change set has a `Modified` and tar has
-        // entry with new content
-
-        fs::write(source_dir.path().join(&b_txt), "Hello")?;
-
-        let snap5 = snap1.repeat();
-        assert_eq!(snap5.entries.len(), 1);
-        assert_eq!(snap5.entries[&b_txt].fingerprint, Some(3297469917561599766));
-
-        let changes = snap4.diff(&snap5);
-        assert_eq!(changes.len(), 1);
-        assert_eq!(changes.items[0], Change::Modified(b_txt.clone()));
-
-        let (.., descriptor) = changes.write_layer(&image_dir, &MediaType::ImageLayerGzip)?;
-        let mut archive = ChangeSet::read_layer(&image_dir, descriptor.digest())?;
-        let mut entries = archive.entries()?;
-        let entry = entries.next().unwrap()?;
-        assert_eq!(entry.path()?, dest_dir.join(b_txt));
-        assert_eq!(entry.size(), 5);
-
-        Ok(())
-    }
-
-    /// Test that the descriptor for a layer is accurate (SHA256 and size are same as
-    /// when independently calculated)
-    #[test]
-    fn changes_layer() -> Result<()> {
-        use std::fs;
-
-        let source_dir = tempdir()?;
-        let image_dir = tempdir()?;
-
-        let snap = Snapshot::new(&source_dir, "workspace");
-
-        fs::write(&source_dir.path().join("some-file.txt"), "Hello")?;
-
-        // Create a layer archive, diffid and descriptor
-
-        let changes = snap.changes();
-        let (diff_id, descriptor) = changes.write_layer(&image_dir, &MediaType::ImageLayerGzip)?;
-
-        assert_eq!(diff_id.len(), 7 + 64);
-        assert!(diff_id.starts_with("sha256:"));
-
-        // Test that size and digest in the descriptor is as for the file
-        let archive = image_dir
-            .path()
-            .join("blobs")
-            .join("sha256")
-            .join(descriptor.digest().strip_prefix("sha256:").unwrap());
-
-        let size = fs::metadata(&archive)?.len() as i64;
-        assert_eq!(descriptor.size(), size);
-
-        let digest = format!("sha256:{}", file_sha256_hex(&archive)?);
-        assert_eq!(descriptor.digest(), &digest);
-
-        Ok(())
-    }
-
     /// Test that when an image is written to a directory that they directory conforms to
     /// the OCI Image Layout spec
     #[tokio::test]
@@ -1636,7 +730,7 @@ mod tests {
         let mut image = Image::new(
             Some(working_dir.path()),
             None,
-            None,
+            Some("ubuntu"),
             None,
             None,
             None,
