+++ rust/images/src/lib.rs
@@ -1,6 +1,9 @@
+pub mod blob_writer;
+pub mod change_set;
 pub mod distribution;
 pub mod image;
 pub mod media_types;
+pub mod snapshot;
 mod utils;
 
 #[cfg(feature = "cli")]
